LPS-70392 Fix LPS-67695 cause issue: place upgrade steps into right upgrade versions
