+++ modules/apps/marketplace/marketplace-service/src/main/java/com/liferay/marketplace/internal/upgrade/MarketplaceServiceUpgrade.java
@@ -34,15 +34,13 @@ public class MarketplaceServiceUpgrade implements UpgradeStepRegistrator {
 	public void register(Registry registry) {
 		registry.register(
 			"com.liferay.marketplace.service", "0.0.1", "2.0.1",
-			new com.liferay.marketplace.internal.upgrade.v0_0_1.UpgradeApp(),
 			new com.liferay.marketplace.internal.upgrade.v0_0_1.UpgradeExpando(
 				_expandoColumnLocalService, _expandoTableLocalService,
-				_expandoValueLocalService),
-			new com.liferay.marketplace.internal.upgrade.v0_0_1.
-				UpgradeModule());
+				_expandoValueLocalService));
 
 		registry.register(
 			"com.liferay.marketplace.service", "1.0.0", "1.0.1",
+			new com.liferay.marketplace.internal.upgrade.v1_0_0.UpgradeApp(),
 			new com.liferay.marketplace.internal.upgrade.v1_0_0.
 				UpgradeModule());
