+++ modules/apps/marketplace/marketplace-service/src/main/java/com/liferay/marketplace/internal/upgrade/v1_0_0/UpgradeModule.java
@@ -37,6 +37,17 @@ public class UpgradeModule extends UpgradeProcess {
 	}
 
 	protected void updateModules() throws Exception {
+		if (!hasColumn("Marketplace_Module", "bundleSymbolicName")) {
+			runSQL(
+				"alter table Marketplace_Module add bundleSymbolicName " +
+					"VARCHAR(500)");
+		}
+
+		if (!hasColumn("Marketplace_Module", "bundleVersion")) {
+			runSQL(
+				"alter table Marketplace_Module add bundleVersion VARCHAR(75)");
+		}
+
 		try (PreparedStatement ps = connection.prepareStatement(
 				"select moduleId, contextName from Marketplace_Module");
 			ResultSet rs = ps.executeQuery()) {
